% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d4t4tur.R
\docType{package}
\name{d4t4tur}
\alias{d4t4tur}
\title{\code{d4t4tur} meta package}
\description{
Caja de Herramientas para el Análisis Turístico de Argentina.
See the README on
\href{https://github.com/dnme-minturdep/d4t4tur/README.md}{Github}
}
