% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{d4t4tur_packages}
\alias{d4t4tur_packages}
\title{List all packages in  d4t4tur}
\usage{
d4t4tur_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include d4t4tur in the list?}
}
\description{
List all packages in  d4t4tur
}
\examples{
d4t4tur_packages()
}
